Customized Product, Add/Remove Customized Product to/from Cart

+++ src/main/java/com/group3/torikago/Torikago/Shop/service/impl/ShoppingCartImplement.java
@@ -4,6 +4,7 @@ import com.group3.torikago.Torikago.Shop.model.CartItems;
 import com.group3.torikago.Torikago.Shop.model.Product;
 import com.group3.torikago.Torikago.Shop.model.User;
 import com.group3.torikago.Torikago.Shop.repository.CartItemRepository;
+import com.group3.torikago.Torikago.Shop.repository.CustomizedBirdCageRepository;
 import com.group3.torikago.Torikago.Shop.repository.ProductRepository;
 import com.group3.torikago.Torikago.Shop.service.ShoppingCartServices;
 import java.util.List;
@@ -15,16 +16,24 @@ import org.springframework.transaction.annotation.Transactional;
 public class ShoppingCartImplement implements ShoppingCartServices{
     private CartItemRepository cartItemRepository;
     private ProductRepository productRepository;
+
+    private CustomizedBirdCageRepository customizedBirdCageRepository;
     
     @Autowired
-    public ShoppingCartImplement(CartItemRepository cartItemRepository, ProductRepository productRepository) {
+    public ShoppingCartImplement(CartItemRepository cartItemRepository, ProductRepository productRepository, CustomizedBirdCageRepository customizedBirdCageRepository) {
         this.cartItemRepository = cartItemRepository;
         this.productRepository = productRepository;
+        this.customizedBirdCageRepository = customizedBirdCageRepository;
     }
 
     @Override
     public List<CartItems> listCartItems(User user) {
          return cartItemRepository.findByUserId(user);
+    }
+
+    @Override
+    public List<CartItems> listCart(User user) {
+        return cartItemRepository.findByUserIdAndCustomizedProductId(user.getId());
     }
 
     @Override
@@ -52,6 +61,12 @@ public class ShoppingCartImplement implements ShoppingCartServices{
 
     @Override
     public void removeProduct(Long productId, User user) {
+        Product product = productRepository.findById(productId);
+        if(product.getCustomizedBirdCage() != null ){
+            product.getCustomizedBirdCage().setCartStatus(false);
+            customizedBirdCageRepository.save(product.getCustomizedBirdCage());
+            return;
+        }
         cartItemRepository.deleteByUserIdAndProductId(user.getId(), productId);
     }
